+++ src/syntax.js
@@ -116,15 +116,15 @@
                     }
 
                     if ( syntaxOptions.highlightComments ) {
-                        innerHTML = renderElementCommentVariables( innerHTML, syntaxLanguage );
-                        innerHTML = renderElementMultiLineCommentVariables( innerHTML, syntaxLanguage );
+                        innerHTML = renderElementCommentVariables( innerHTML, syntaxLanguage, syntaxOptions );
+                        innerHTML = renderElementMultiLineCommentVariables( innerHTML, syntaxLanguage, syntaxOptions );
                     }
 
                     if ( syntaxOptions.highlightStrings ) {
-                        innerHTML = renderElementStringQuotesPatternVariables( innerHTML, innerHTML.match( /".*?"/g ) );
+                        innerHTML = renderElementStringQuotesPatternVariables( innerHTML, innerHTML.match( /".*?"/g ), syntaxOptions );
 
                         if ( _languages[ syntaxLanguage ].comment !== "'" ) {
-                            innerHTML = renderElementStringQuotesPatternVariables( innerHTML, innerHTML.match( /'.*?'/g ) );
+                            innerHTML = renderElementStringQuotesPatternVariables( innerHTML, innerHTML.match( /'.*?'/g ), syntaxOptions );
                         }
                     }
 
@@ -152,7 +152,7 @@
         }
     }
 
-    function renderElementCommentVariables( innerHTML, syntaxLanguage ) {
+    function renderElementCommentVariables( innerHTML, syntaxLanguage, syntaxOptions ) {
         var lookup = _languages[ syntaxLanguage ].comment,
             patternItems = innerHTML.match( new RegExp( lookup + ".*", "g" ) );
 
@@ -167,13 +167,15 @@
                 _comments_Cached_Count++;
     
                 innerHTML = innerHTML.replace( comment, commentVariable );
+
+                fireCustomTrigger( syntaxOptions.onCommentRender, comment );
             }
         }
 
         return innerHTML;
     }
 
-    function renderElementMultiLineCommentVariables( innerHTML, syntaxLanguage ) {
+    function renderElementMultiLineCommentVariables( innerHTML, syntaxLanguage, syntaxOptions ) {
         var lookup = _languages[ syntaxLanguage ].multiLineComment;
 
         if ( isDefinedArray( lookup ) && lookup.length === 2 ) {
@@ -199,8 +201,9 @@
                             _comments_Cached_Count++;
                 
                             innerHTML = innerHTML.replace( commentLine, commentVariable );
-    
                         }
+
+                        fireCustomTrigger( syntaxOptions.onCommentRender, comment );
                     }
                 }
             }
@@ -209,7 +212,7 @@
         return innerHTML;
     }
 
-    function renderElementStringQuotesPatternVariables( innerHTML, patternItems ) {
+    function renderElementStringQuotesPatternVariables( innerHTML, patternItems, syntaxOptions ) {
         if ( patternItems !== null ) {
             var patternItemsLength = patternItems.length;
         
@@ -222,6 +225,8 @@
                 _strings_Cached_Count++;
     
                 innerHTML = innerHTML.replace( quote, quoteVariable );
+
+                fireCustomTrigger( syntaxOptions.onStringRender, quote );
             }
         }
 
@@ -243,6 +248,8 @@
             } else {
                 innerHTML = innerHTML.replace( regEx, "<span class=\"keyword\">" + keyword + "</span>" );
             }
+
+            fireCustomTrigger( syntaxOptions.onKeywordRender, keyword );
         }
 
         return innerHTML;
@@ -364,6 +371,9 @@
         options.onCopy = getDefaultFunction( options.onCopy, null );
         options.onRenderComplete = getDefaultFunction( options.onRenderComplete, null );
         options.onKeywordClicked = getDefaultFunction( options.onKeywordClicked, null );
+        options.onKeywordRender = getDefaultFunction( options.onKeywordRender, null );
+        options.onStringRender = getDefaultFunction( options.onStringRender, null );
+        options.onCommentRender = getDefaultFunction( options.onCommentRender, null );
 
         return options;
     }