+++ docs/CHANGE_LOG.md
@@ -5,6 +5,9 @@
 - The render() method now only scans for "div" and "code" DOM element types (faster).
 - Improved the README.md and README_NUGET.md documentation.
 - When a specific language is not supported, a new error log is shown in the console.
+- Added new custom trigger "onKeywordRender", which is called when a keyword is rendered (before main render completion).
+- Added new custom trigger "onStringRender", which is called when a string is rendered (before main render completion).
+- Added new custom trigger "onCommentRender", which is called when a comment is rendered (before main render completion).
 - 
 
 <br>